requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.1
cron-descriptor==1.4.5
cycler==0.12.1
Django==5.1.8
django-celery-beat==2.7.0
django-timezone-field==7.1
djangorestframework==3.16.0
fonttools==4.57.0
kiwisolver==1.4.8
kombu==5.5.2
matplotlib==3.10.1
narwhals==1.33.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.1
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pyparsing==3.2.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
